+++ clients/admin-ui/cypress/e2e/systems-plus.cy.ts
@@ -491,8 +491,7 @@ describe("System management with Plus features", () => {
       cy.wait("@getSystemAssets");
       cy.getByTestId("row-0-col-name").should("contain", "ar_debug");
       cy.getByTestId("row-0-col-locations").should("contain", "United States");
-      cy.getByTestId("row-0-col-data_uses").should("contain", "Analytics");
-      cy.getByTestId("row-0-col-data_uses").should("not.contain", "employment");
+      cy.getByTestId("row-0-col-data_uses").should("contain", "Essential");
     });
 
     describe("asset operations", () => {
@@ -508,7 +507,7 @@ describe("System management with Plus features", () => {
         cy.getByTestId("input-domain").type("example.com");
         cy.getByTestId("controlled-select-asset_type").antSelect("Cookie");
         cy.getByTestId("controlled-select-data_uses").antSelect("analytics");
-        cy.getByTestId("save-btn").click();
+        cy.getByTestId("save-btn").click({ force: true });
         cy.wait("@addSystemAsset");
       });
 
@@ -524,10 +523,11 @@ describe("System management with Plus features", () => {
         cy.getByTestId("controlled-select-asset_type")
           .should("contain", "Cookie")
           .should("have.class", "ant-select-disabled");
-        cy.getByTestId("controlled-select-data_uses")
-          .should("contain", "analytics")
-          .should("not.contain", "employment")
-          .antSelect("essential");
+        cy.getByTestId("controlled-select-data_uses").should(
+          "contain",
+          "essential",
+        );
+        cy.getByTestId("controlled-select-data_uses").antSelect(0);
         cy.getByTestId("input-domain")
           .should("have.value", ".doubleclick.net")
           .should("be.disabled");
@@ -539,9 +539,8 @@ describe("System management with Plus features", () => {
         cy.getByTestId("save-btn").click();
         cy.wait("@updateSystemAssets").then((interception) => {
           expect(interception.request.body[0].data_uses).to.eql([
-            "analytics",
             "essential",
-            "employment",
+            "analytics",
           ]);
         });
       });
@@ -552,7 +551,7 @@ describe("System management with Plus features", () => {
         });
 
         cy.getByTestId("confirmation-modal").should("exist");
-        cy.getByTestId("continue-btn").click();
+        cy.getByTestId("continue-btn").click({ force: true });
         cy.wait("@deleteSystemAssets");
       });
 
@@ -566,6 +565,10 @@ describe("System management with Plus features", () => {
           "Javascript tag",
         );
         cy.getByTestId("controlled-select-data_uses").antSelect("analytics");
+        cy.getByTestId("controlled-select-data_uses").within(() => {
+          // force select menu to close so it doesn't cover the input
+          cy.get("input").focus().blur();
+        });
         // blur the input without entering anything to trigger the error
         cy.getByTestId("input-base_url").clear().blur();
         cy.getByTestId("save-btn").should("be.disabled");