+++ clients/admin-ui/src/features/datastore-connections/add-connection/forms/ConnectorParametersForm.tsx
@@ -7,7 +7,6 @@ import {
 import { useLazyGetDatastoreConnectionStatusQuery } from "datastore-connections/datastore-connection.slice";
 import {
   AntButton as Button,
-  CircleHelpIcon,
   Flex,
   FormControl,
   FormErrorMessage,
@@ -20,13 +19,13 @@ import {
   NumberInputField,
   NumberInputStepper,
   Textarea,
-  Tooltip,
   VStack,
 } from "fidesui";
 import { Field, FieldInputProps, Form, Formik, FormikProps } from "formik";
 import React, { useEffect, useRef } from "react";
 
 import { useAppSelector } from "~/app/hooks";
+import { InfoTooltip } from "~/features/common/InfoTooltip";
 import { ConnectionType } from "~/types/api";
 
 import {
@@ -172,21 +171,9 @@ const ConnectorParametersForm = ({
             )}
             <FormErrorMessage>{form.errors[key]}</FormErrorMessage>
           </VStack>
-          <Tooltip
-            aria-label={item.description}
-            hasArrow
-            label={item.description}
-            placement="right-start"
-            openDelay={500}
-          >
-            <Flex
-              alignItems="center"
-              h="32px"
-              visibility={item.description ? "visible" : "hidden"}
-            >
-              <CircleHelpIcon marginLeft="8px" _hover={{ cursor: "pointer" }} />
-            </Flex>
-          </Tooltip>
+          <Flex alignItems="center" h={8} ml={2}>
+            <InfoTooltip label={item.description} />
+          </Flex>
         </FormControl>
       )}
     </Field>
@@ -265,7 +252,7 @@ const ConnectorParametersForm = ({
                   isInvalid={props.errors.name && props.touched.name}
                 >
                   {getFormLabel("name", "Name")}
-                  <VStack align="flex-start" w="inherit">
+                  <VStack align="flex-start" w="inherit" mr="6">
                     <Input
                       {...field}
                       autoComplete="off"
@@ -279,9 +266,6 @@ const ConnectorParametersForm = ({
                     />
                     <FormErrorMessage>{props.errors.name}</FormErrorMessage>
                   </VStack>
-                  <Flex alignItems="center" h="32px" visibility="hidden">
-                    <CircleHelpIcon marginLeft="8px" />
-                  </Flex>
                 </FormControl>
               )}
             </Field>
@@ -299,10 +283,8 @@ const ConnectorParametersForm = ({
                     resize="none"
                     size="sm"
                     value={field.value || ""}
+                    mr="6"
                   />
-                  <Flex alignItems="center" h="32px" visibility="hidden">
-                    <CircleHelpIcon marginLeft="8px" />
-                  </Flex>
                 </FormControl>
               )}
             </Field>
@@ -336,20 +318,9 @@ const ConnectorParametersForm = ({
                       {props.errors.instance_key}
                     </FormErrorMessage>
                   </VStack>
-                  <Tooltip
-                    aria-label="The fides_key will allow fidesops to associate dataset field references appropriately. Must be a unique alphanumeric value with no spaces (underscores allowed) to represent this connection."
-                    hasArrow
-                    label="The fides_key will allow fidesops to associate dataset field references appropriately. Must be a unique alphanumeric value with no spaces (underscores allowed) to represent this connection."
-                    placement="right-start"
-                    openDelay={500}
-                  >
-                    <Flex alignItems="center" h="32px">
-                      <CircleHelpIcon
-                        marginLeft="8px"
-                        _hover={{ cursor: "pointer" }}
-                      />
-                    </Flex>
-                  </Tooltip>
+                  <Flex alignItems="center" h={8} ml={2}>
+                    <InfoTooltip label="The fides_key will allow fidesops to associate dataset field references appropriately. Must be a unique alphanumeric value with no spaces (underscores allowed) to represent this connection." />
+                  </Flex>
                 </FormControl>
               )}
             </Field>
@@ -361,7 +332,7 @@ const ConnectorParametersForm = ({
               }
               return getFormField(key, item);
             })}
-            <div className="flex gap-2">
+            <Flex gap={2}>
               <Button
                 disabled={!connection?.key}
                 loading={result.isLoading || result.isFetching}
@@ -378,7 +349,7 @@ const ConnectorParametersForm = ({
               >
                 Save
               </Button>
-            </div>
+            </Flex>
           </VStack>
         </Form>
       )}