+++ clients/admin-ui/src/features/datastore-connections/system_portal_config/forms/ConnectorParametersForm.tsx
@@ -8,7 +8,6 @@ import DSRCustomizationModal from "datastore-connections/system_portal_config/fo
 import {
   AntButton as Button,
   AntSelect as Select,
-  CircleHelpIcon,
   Flex,
   FormControl,
   FormErrorMessage,
@@ -20,7 +19,6 @@ import {
   NumberInputField,
   NumberInputStepper,
   Spacer,
-  Tooltip,
   VStack,
 } from "fidesui";
 import { Field, FieldInputProps, Form, Formik, FormikProps } from "formik";
@@ -29,6 +27,7 @@ import React from "react";
 import { DatastoreConnectionStatus } from "src/features/datastore-connections/types";
 
 import { useFeatures } from "~/features/common/features";
+import { InfoTooltip } from "~/features/common/InfoTooltip";
 import DisableConnectionModal from "~/features/datastore-connections/DisableConnectionModal";
 import SelectDataset from "~/features/datastore-connections/system_portal_config/forms/SelectDataset";
 import {
@@ -217,24 +216,9 @@ export const ConnectorParametersForm = ({
               )}
               <FormErrorMessage>{error}</FormErrorMessage>
             </VStack>
-            <Tooltip
-              aria-label={item.description}
-              hasArrow
-              label={item.description}
-              placement="right-start"
-              openDelay={500}
-            >
-              <Flex
-                alignItems="center"
-                h="32px"
-                visibility={item.description ? "visible" : "hidden"}
-              >
-                <CircleHelpIcon
-                  marginLeft="8px"
-                  _hover={{ cursor: "pointer" }}
-                />
-              </Flex>
-            </Tooltip>
+            <Flex alignItems="center" h={8} ml={2}>
+              <InfoTooltip label={item.description} />
+            </Flex>
           </FormControl>
         );
       }}
@@ -403,20 +387,9 @@ export const ConnectorParametersForm = ({
                           {props.errors.instance_key as string}
                         </FormErrorMessage>
                       </VStack>
-                      <Tooltip
-                        aria-label="The fides_key will allow fidesops to associate dataset field references appropriately. Must be a unique alphanumeric value with no spaces (underscores allowed) to represent this integration."
-                        hasArrow
-                        label="The fides_key will allow fidesops to associate dataset field references appropriately. Must be a unique alphanumeric value with no spaces (underscores allowed) to represent this integration."
-                        placement="right-start"
-                        openDelay={500}
-                      >
-                        <Flex alignItems="center" h="32px">
-                          <CircleHelpIcon
-                            marginLeft="8px"
-                            _hover={{ cursor: "pointer" }}
-                          />
-                        </Flex>
-                      </Tooltip>
+                      <Flex alignItems="center" h={8} ml={2}>
+                        <InfoTooltip label="The fides_key will allow fidesops to associate dataset field references appropriately. Must be a unique alphanumeric value with no spaces (underscores allowed) to represent this integration." />
+                      </Flex>
                     </FormControl>
                   )}
                 </Field>
@@ -486,20 +459,9 @@ export const ConnectorParametersForm = ({
                           {props.errors.enabled_actions as string}
                         </FormErrorMessage>
                       </VStack>
-                      <Tooltip
-                        aria-label="The request types that are supported for this integration."
-                        hasArrow
-                        label="The request types that are supported for this integration."
-                        placement="right-start"
-                        openDelay={500}
-                      >
-                        <Flex alignItems="center" h="32px">
-                          <CircleHelpIcon
-                            marginLeft="8px"
-                            _hover={{ cursor: "pointer" }}
-                          />
-                        </Flex>
-                      </Tooltip>
+                      <Flex alignItems="center" h={8} ml={2}>
+                        <InfoTooltip label="The request types that are supported for this integration." />
+                      </Flex>
                     </FormControl>
                   )}
                 </Field>